requirements.txt
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
openai-whisper
torch
ffmpeg-python
pytest
python-dotenv